+++ UnityProject/Assets/Resources/ScriptableObjectsSingletons/SpriteCatalogueSingleton.asset
@@ -10416,6 +10416,7 @@ MonoBehaviour:
   - {fileID: 11400000, guid: e3926ff273f522f4e93f5770bb1df908, type: 2}
   - {fileID: 11400000, guid: cca8b419f51676744827801194f1eb78, type: 2}
   - {fileID: 11400000, guid: f92d0aacc426db9428d37a3aada55b7d, type: 2}
+  - {fileID: 11400000, guid: 13e6a8c51327f4e4b8cf017df19e6c13, type: 2}
   - {fileID: 11400000, guid: d3618c58d1706d946923dd52e5fb6dd5, type: 2}
   - {fileID: 11400000, guid: 6fc4547c2b620f84486e7dac49b114d5, type: 2}
   - {fileID: 11400000, guid: 96a177d9d65cd61439a9f8f1177a5def, type: 2}
@@ -17649,6 +17650,14 @@ MonoBehaviour:
   - {fileID: 11400000, guid: c8aa0f4667f4ac44a957cafcf77bf266, type: 2}
   - {fileID: 11400000, guid: 3e7793534207ccb4f8c0baf38313cf59, type: 2}
   - {fileID: 11400000, guid: 4d7348ae0e62f2b4ebe0f86e52797823, type: 2}
+  - {fileID: 11400000, guid: 50460fd44e949d2458d1020eae197cdd, type: 2}
+  - {fileID: 11400000, guid: 4b8905b04fb6545468645a629173092c, type: 2}
+  - {fileID: 11400000, guid: e385360b8ed26e44ba3a3c5b2b389add, type: 2}
+  - {fileID: 11400000, guid: 94b4ef2504964e94cb404f0d0aa67e25, type: 2}
+  - {fileID: 11400000, guid: 0ba208bd2ec4e8c4c938c188a570d2b5, type: 2}
+  - {fileID: 11400000, guid: 4f172c381dca8fe40adc2c781edb035a, type: 2}
+  - {fileID: 11400000, guid: 2d96a0de87f852e40958ddf0a94a30a9, type: 2}
+  - {fileID: 11400000, guid: b47b939d4743b124188ff4f58517fff0, type: 2}
   - {fileID: 11400000, guid: c5d856adabdbffc44acca9eb60f114f2, type: 2}
   - {fileID: 11400000, guid: 63c2ebaf816ee7d4c817e53f446748f3, type: 2}
   - {fileID: 11400000, guid: 547a8be3ca3301d40a75f8c13c207626, type: 2}
@@ -17692,8 +17701,6 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 77f03a126f8b5354ea022b4b54340723, type: 2}
   - {fileID: 11400000, guid: ac3e76dc2d169174fa66e5159afeb621, type: 2}
   - {fileID: 11400000, guid: d47c52715a1cfe34c86031c75bac33d0, type: 2}
-  - {fileID: 11400000, guid: e385360b8ed26e44ba3a3c5b2b389add, type: 2}
-  - {fileID: 11400000, guid: 94b4ef2504964e94cb404f0d0aa67e25, type: 2}
   - {fileID: 11400000, guid: 63322873d6cd2a24380f89d6bc7a5bb5, type: 2}
   - {fileID: 11400000, guid: 5b60b4e4267fef54ca2c8f940d26e798, type: 2}
   - {fileID: 11400000, guid: d7aa5607a1def354aadab44ad046d752, type: 2}
@@ -17770,8 +17777,6 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 3144410c8f2940c488d9a7f8d07673ea, type: 2}
   - {fileID: 11400000, guid: 3702b3c7ba3242f4488e6c5f5ce5cf5b, type: 2}
   - {fileID: 11400000, guid: 4e599cbc0267f074e94e32e95e78914f, type: 2}
-  - {fileID: 11400000, guid: 2d96a0de87f852e40958ddf0a94a30a9, type: 2}
-  - {fileID: 11400000, guid: b47b939d4743b124188ff4f58517fff0, type: 2}
   - {fileID: 11400000, guid: 1f3ae7e3a112c804fbd7663d160eff74, type: 2}
   - {fileID: 11400000, guid: 6c2ba822f5a7687449f8326ab14e41e5, type: 2}
   - {fileID: 11400000, guid: 14e36c88027c93a428d7c344a4eb891a, type: 2}
@@ -18641,6 +18646,16 @@ MonoBehaviour:
   - {fileID: 11400000, guid: f97f85e6099a0a14e9531428fe35a6bd, type: 2}
   - {fileID: 11400000, guid: 79f2fb8f304970640830af78961a274c, type: 2}
   - {fileID: 11400000, guid: 55b896508c1a47146b0abcf1ba92fee5, type: 2}
+  - {fileID: 11400000, guid: cbed5f9c081a08246a20f157d6b7396d, type: 2}
+  - {fileID: 11400000, guid: 669132097c1c9cc4faa95f2571d09ab9, type: 2}
+  - {fileID: 11400000, guid: a8eb9ff8c0cc22547bf3ff737d7080de, type: 2}
+  - {fileID: 11400000, guid: 27943fd3d65870040b566dc1134b1062, type: 2}
+  - {fileID: 11400000, guid: 24836d0bac0d9f444915764673b40863, type: 2}
+  - {fileID: 11400000, guid: 1222c962f0dce0340ad3866839eaba36, type: 2}
+  - {fileID: 11400000, guid: 8dfa079b7b1833e4fbec24ccb22222d5, type: 2}
+  - {fileID: 11400000, guid: 26ad6140762fc61448b727c584089451, type: 2}
+  - {fileID: 11400000, guid: 0997f3514c83e894399b2d10ecf7acc1, type: 2}
+  - {fileID: 11400000, guid: 084406690e001df408b793c70e4f1a3f, type: 2}
   - {fileID: 11400000, guid: 07d40c6f24d8d4f45afd6173f413fc2d, type: 2}
   - {fileID: 11400000, guid: 5bb15d6fbaeb2a34f81da2e9a651058c, type: 2}
   - {fileID: 11400000, guid: ce47a1913229b744eb72d75a7e1675ac, type: 2}
@@ -24028,4 +24043,3 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 1f9843f716e40fb4187e3ecde15d87b3, type: 2}
   - {fileID: 11400000, guid: 01bd1367ae0668b48b7dfef3135c9180, type: 2}
   - {fileID: 11400000, guid: 58abfc11d31324549a8595d74e115da8, type: 2}
-  - {fileID: 11400000, guid: 3b618664e162a074880e5e086ccd1578, type: 2}
