+++ UnityProject/Assets/Scripts/UI/Objects/Research/BlastYieldDetector/GUI_BlastYieldDetector.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using UI.Core.NetUI;
+﻿using UI.Core.NetUI;
 using UnityEngine;
 using Systems.Research.Objects;
 using Systems.Research;
@@ -25,13 +24,7 @@ namespace UI.Objects.Research
 			}
 		}
 
-		private const int YAXIS_MAX = 1000; //Up to 1000 Blast yield
-		private const int XAXIS_MAX = 10; //Up to last 10 datapoints
-
 		private BlastYieldDetector _blastYieldDetector;
-
-		private GUI_BlastYieldDetector clientGUI;
-		private Transform clientGUIGraphTransform;
 
 		private List<ExplosiveBounty> bountyList;
 
@@ -60,34 +53,24 @@ namespace UI.Objects.Research
 		private EmptyItemList bountyContainer;
 
 		[SerializeField]
-		public EmptyItemList graphContainer;
-
-		/// <summary>
-		/// Offset to position highlight line UI properly
-		/// </summary>
-		public float rectOffset;
-		[SerializeField]
-		private RectTransform yieldNodeHighlight;
-
-		[SerializeField]
-		private RectTransform pointNodeHighlight;
+		private GUI_BYDGraph blastYieldGraph;
 
 		#endregion
 
 		#region Initialization
 		private void Start()
 		{
-			clientGUI = UIManager.Instance.transform.GetChild(0).GetComponentInChildren<GUI_BlastYieldDetector>();
-			clientGUIGraphTransform = clientGUI.graphContainer.transform;
 			BlastYieldDetector.blastEvent += OnRecieveBlast;
 			BlastYieldDetector.updateGUIEvent += UpdateGui;
-			UpdateDataDisplay();
+
 			UpdateGui();
+
+			blastYieldGraph.UpdateDataDisplay(blastYieldDetector);
 		}
 		#endregion
 
 		private void OnRecieveBlast(BlastData data)
-		{
+		{  
 			var mix = data.ReagentMix;
 			if (smokeReaction.IsReactionValid(mix)) smokeLabel.MasterSetValue(smokeReaction.GetReactionAmount(mix).ToString());
 			else smokeLabel.MasterSetValue("0");
@@ -98,7 +81,7 @@ namespace UI.Objects.Research
 			reagentLabel.MasterSetValue(data.ReagentMix.Total.ToString());
 			yieldLabel.MasterSetValue(blastYieldDetector.BlastYieldData[blastYieldDetector.BlastYieldData.Count - 1].ToString());
 
-			UpdateDataDisplay();
+			blastYieldGraph.UpdateDataDisplay(blastYieldDetector);
 
 			UpdateGui();
 		}
@@ -135,6 +118,10 @@ namespace UI.Objects.Research
 				entryScript.Initialise(bounties[i], i);
 			}
 		}
+		public void SetCurrentShownData(int pos)
+		{
+			blastYieldGraph.UpdateDataDisplay(blastYieldDetector);
+		}
 
 		public void OnDestroy()
 		{
@@ -142,67 +129,6 @@ namespace UI.Objects.Research
 			BlastYieldDetector.updateGUIEvent -= UpdateGui;
 		}
 
-		#region Plotting
-
-		public Vector2 GetNodePosition(float yield, float index)
-		{
-			float yieldClamp = Math.Min(yield, YAXIS_MAX);
-
-			float dotPosY = yieldClamp * graphContainer.GetComponent<RectTransform>().rect.height / YAXIS_MAX;
-
-			//points axis position
-			float dotPosX = index * graphContainer.GetComponent<RectTransform>().rect.width / XAXIS_MAX;
-
-			//position 2d, third axis isn't important
-			Vector2 dotPosition = new Vector2(dotPosX, dotPosY);
-			return dotPosition;
-		}
-
-		public void SetCurrentShownData(int pos)
-		{
-			UpdateDataDisplay();
-		}
-
-		/// <summary>
-		/// Moves highlight lines to current node, and updates labels
-		/// </summary>
-		private void UpdateDataDisplay()
-		{
-			if (blastYieldDetector == null || blastYieldDetector.BlastYieldData == null)
-			{
-				graphContainer.Clear();
-				return;
-			}
-
-			List<float> yields = blastYieldDetector.BlastYieldData.ToList();
-			if (yields.Count > XAXIS_MAX)
-			{
-				yields = yields.GetRange(blastYieldDetector.BlastYieldData.Count - 1 - XAXIS_MAX, XAXIS_MAX); //Obtains last ten datapoints
-			}
-
-			graphContainer.SetItems(yields.Count);
-
-			for (int i = 0; i < yields.Count; i++)
-			{
-				Vector2 dataShownPos = GetNodePosition(yields[i], i);
-
-				if (i < clientGUIGraphTransform.childCount)
-				{
-					clientGUIGraphTransform.GetChild(i).GetComponent<RectTransform>().anchoredPosition = dataShownPos;
-				}
-
-				if (i != yields.Count - 1) continue;
-
-				Vector3 yieldNewY = yieldNodeHighlight.anchoredPosition;
-				yieldNewY.y = dataShownPos.y + rectOffset;
-				clientGUI.yieldNodeHighlight.anchoredPosition = yieldNewY;
-
-				Vector3 pointNewX = pointNodeHighlight.anchoredPosition;
-				pointNewX.x = dataShownPos.x + rectOffset;
-				clientGUI.pointNodeHighlight.anchoredPosition = pointNewX;
-			}
-		}
-
-		#endregion
+		
 	}
 }
